+++ sphinx_needs/needs.py
@@ -247,6 +247,9 @@ def setup(app: Sphinx) -> Dict[str, Any]:
     app.add_config_value("needs_variants", {}, "html", types=[dict])
     app.add_config_value("needs_variant_options", [], "html", types=[list])
 
+    # add jinja context option
+    app.add_config_value("needs_render_context", {}, "html", types=[dict])
+
     # Define nodes
     app.add_node(Need, html=(html_visit, html_depart), latex=(latex_visit, latex_depart))
     app.add_node(