+++ sphinx_needs/services/open_needs.py
@@ -111,7 +111,8 @@ class OpenNeedsService(BaseService):
         needs_id_validator = self.app.config.needs_id_regex or "^[A-Z0-9_]{5,}"
         for item in data:
             if isinstance(id_selector, str):
-                value = jinja_parse(item, id_selector)
+                context = {**item, **self.app.config.needs_render_context}
+                value = jinja_parse(context, id_selector)
             else:
                 value = str(dict_get(item, id_selector))
             # Validate the value for a need ID or generate a valid ID for the need
@@ -133,14 +134,15 @@ class OpenNeedsService(BaseService):
                 if isinstance(selector, str):
                     # Set the "hard-coded" string or
                     # combine the "hard-coded" string and dynamic value
-                    selector = jinja_parse(item, selector)
+                    context = {**item, **self.app.config.needs_render_context}
+                    selector = jinja_parse(context, selector)
                     # Set the returned string as value
                     extra_data[name] = selector
                 else:
                     extra_data[name] = dict_get(item, selector)
 
             content_template = Template(self.content, autoescape=True)
-            context = {"data": item, "options": options}
+            context = {"data": item, "options": options, **self.app.config.needs_render_context}
             content = content_template.render(context)
             content += "\n\n| \n"  # Add enough space between content and extra_data
 
@@ -162,7 +164,8 @@ class OpenNeedsService(BaseService):
                 if isinstance(selector, str):
                     # Set the "hard-coded" string or
                     # combine the "hard-coded" string and dynamic value
-                    selector = jinja_parse(item, selector)
+                    context = {**item, **self.app.config.needs_render_context}
+                    selector = jinja_parse(context, selector)
                     # Set the returned string as value
                     need_values[name] = selector
                 else: