+++ docs/changelog.rst
@@ -13,12 +13,14 @@ License
 Under development
 
 * Improvement: Supported arguments as filter string for :ref:`needextract`.
-  (`#688 <https://github.com/useblocks/sphinx-needs/issues/688>`_)s
+  (`#688 <https://github.com/useblocks/sphinx-needs/issues/688>`_)
+* Improvement: Added :ref:`needs_render_context` configuration option which enables you to use custom data as the
+context when rendering Jinja templates or strings.
+  (`#704 <https://github.com/useblocks/sphinx-needs/issues/704>`_)
 * Improvement: Supported ``target_url`` for :ref:`needs_external_needs`.
   (`#701 <https://github.com/useblocks/sphinx-needs/issues/701>`_)
 * Bugfix: Fixed needuml key shown in need meta data by providing internal need option `arch`.
   (`#687 <https://github.com/useblocks/sphinx-needs/issues/687>`_)
-
 
 1.0.2
 -----